# Improve performance of the bitwarden lookup plugin