--- a/changelogs/fragments/bitwarden-lookup-performance.yaml
+++ b/changelogs/fragments/bitwarden-lookup-performance.yaml
@@ -1,3 +1,2 @@
 minor_changes:
-  - bitwarden lookup plugin: When looking for items using an item id, the item is now accessed directly with bw get item instead of searching through all items. 
-This doubles the lookup speed.
+  - "bitwarden lookup plugin - when looking for items using an item ID, the item is now accessed directly with ``bw get item`` instead of searching through all items. This doubles the lookup speed (https://github.com/ansible-collections/community.general/pull/6619)."