When looking for items using an item id, we can access the item directly with bw get item instead of searching through all items. This doubles the lookup speed.

##### SUMMARY
<!--- Describe the change below, including rationale and design decisions -->

<!--- HINT: Include "Fixes #nnn" if you are fixing an existing issue -->

<!--- Please do not forget to include a changelog fragment:
      https://docs.ansible.com/ansible/devel/community/collection_development_process.html#creating-changelog-fragments
      No need to include one for docs-only or test-only PR, and for new plugin/module PRs.
      Read about more details in CONTRIBUTING.md.
      -->
The bitwarden lookup plugin searches through all items using `bw list items --search`, which takes a lot of time. However, when looking for an item using the item id, there is a much faster way: `bw get item  <itemid>`. This PR introduces this method for every lookup that uses the item's id.

Some comparisons:
```
bw list items --search 52c0e9ce***  1,31s user 0,14s system 126% cpu 1,150 total
bw get item 52c0e9ce***  0,59s user 0,10s system 105% cpu 0,655 total
```

Or using a task that runs twelve lookups against a single host:
Before this change: 46.52s
After this change: 25.65s

This is close to double speed.

##### ISSUE TYPE
<!--- Pick one or more below and delete the rest.
      'Test Pull Request' is for PRs that add/extend tests without code changes. -->
- Feature/Performance Pull Request

##### COMPONENT NAME
<!--- Write the SHORT NAME of the module, plugin, task or feature below. -->
bitwarden
##### ADDITIONAL INFORMATION
<!--- Include additional information to help people understand the change here -->
<!--- A step-by-step reproduction of the problem is helpful if there is no related issue -->
As `bw list` returns a list, but `bw get` a single item, I choose to put it back into a list to not introduce any breaking changes here.
<!--- Paste verbatim command output below, e.g. before and after your change -->
```paste below

```