@felixfontein 2023-07-15
What's the current state on https://github.com/ansible-collections/community.general/pull/6619#discussion_r1218495894?

needs_info